add: draw string utils

+++ src/layout/lane.c
@@ -8,55 +8,55 @@ void LaneLineRender(int laneCount) {
     int screenWidth = GetScreenWidth();
     int screenHeight = GetScreenHeight();
 
-    float lineWidth = (float) screenWidth * LANE_LINE_WIDTH;
+    int lineWidth = screenWidth * LANE_LINE_WIDTH;
 
-    float noteWidth = (float) screenWidth * LANE_NOTE_WIDTH;
-    float noteHeight = (float) screenHeight * LANE_NOTE_HEIGHT;
+    int noteWidth = screenWidth * LANE_NOTE_WIDTH;
+    int noteHeight = screenHeight * LANE_NOTE_HEIGHT;
 
-    float infoXPos = (float) screenWidth * INFO_GAP;
-    float infoWidth = (float) screenWidth * INFO_WIDTH;
+    int infoXPos = screenWidth * INFO_GAP;
+    int infoWidth = screenWidth * INFO_WIDTH;
 
-    float fieldWidth = (float) noteWidth * laneCount;
-    float fieldXPos = infoWidth + infoXPos * 2 + lineWidth;
+    int fieldWidth = noteWidth * laneCount;
+    int fieldXPos = infoWidth + infoXPos * 2 + lineWidth;
 
-    float judgmentYPos = (float) screenHeight - ((float) screenHeight * LANE_JUDGEMENT_Y);
+    int judgmentYPos = screenHeight - (screenHeight * LANE_JUDGEMENT_Y);
 
     DrawRectangle(
-        (int) fieldXPos,
+        fieldXPos,
         0,
-        (int) fieldWidth,
-        (int) judgmentYPos - noteHeight / 2,
+        fieldWidth,
+        judgmentYPos - noteHeight / 2,
         LANE_BACKGROUND_COLOR
     );
 
     DrawRectangle(
-        (int) fieldXPos,
-        (int) judgmentYPos - noteHeight / 2,
-        (int) fieldWidth,
-        (int) noteHeight,
+        fieldXPos,
+        judgmentYPos - noteHeight / 2,
+        fieldWidth,
+        noteHeight,
         LANE_JUDGEMENT_COLOR
     );
 
     DrawRectangle(
-        (int) fieldXPos,
-        (int) judgmentYPos + noteHeight / 2,
-        (int) fieldWidth,
-        (int) screenHeight,
+        fieldXPos,
+        judgmentYPos + noteHeight / 2,
+        fieldWidth,
+        screenHeight,
         LANE_BACKGROUND_COLOR
     );
 
     DrawRectangle(
-        (int) fieldXPos - lineWidth,
+        fieldXPos - lineWidth,
         0,
-        (int) lineWidth,
+        lineWidth,
         screenHeight,
         LANE_LINE_COLOR
     );
 
     DrawRectangle(
-        (int) fieldXPos + fieldWidth,
+        fieldXPos + fieldWidth,
         0,
-        (int) lineWidth,
+        lineWidth,
         screenHeight,
         LANE_LINE_COLOR
     );
@@ -66,20 +66,20 @@ void LaneKeyPressRender(int laneCount) {
     int screenWidth = GetScreenWidth();
     int screenHeight = GetScreenHeight();
 
-    float noteWidth = (float) screenWidth * LANE_NOTE_WIDTH;
+    int noteWidth = screenWidth * LANE_NOTE_WIDTH;
 
-    float infoXPos = (float) screenWidth * INFO_GAP;
-    float infoWidth = (float) screenWidth * INFO_WIDTH;
+    int infoXPos = screenWidth * INFO_GAP;
+    int infoWidth = screenWidth * INFO_WIDTH;
 
-    float fieldXPos = infoWidth + infoXPos * 2;
+    int fieldXPos = infoWidth + infoXPos * 2;
 
     switch (laneCount) {
         case 4:
             if (IsKeyDown(LANE_4K_1)) {
                 DrawRectangle(
-                    (int) fieldXPos,
+                    fieldXPos,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -87,9 +87,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_4K_2)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth,
+                    fieldXPos + noteWidth,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -97,9 +97,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_4K_3)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 2,
+                    fieldXPos + noteWidth * 2,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -107,9 +107,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_4K_4)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 3,
+                    fieldXPos + noteWidth * 3,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -119,9 +119,9 @@ void LaneKeyPressRender(int laneCount) {
         case 5:
             if (IsKeyDown(LANE_5K_1)) {
                 DrawRectangle(
-                    (int) fieldXPos,
+                    fieldXPos,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -129,9 +129,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_5K_2)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth,
+                    fieldXPos + noteWidth,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -139,17 +139,17 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_5K_3)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 2,
+                    fieldXPos + noteWidth * 2,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
             } else if (IsKeyDown(LANE_5K_4)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 2,
+                    fieldXPos + noteWidth * 2,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -157,9 +157,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_5K_5)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 3,
+                    fieldXPos + noteWidth * 3,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -167,9 +167,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_5K_6)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 4,
+                    fieldXPos + noteWidth * 4,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -179,9 +179,9 @@ void LaneKeyPressRender(int laneCount) {
         case 6:
             if (IsKeyDown(LANE_6K_1)) {
                 DrawRectangle(
-                    (int) fieldXPos,
+                    fieldXPos,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -189,9 +189,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_6K_2)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth,
+                    fieldXPos + noteWidth,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -199,9 +199,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_6K_3)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 2,
+                    fieldXPos + noteWidth * 2,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -209,9 +209,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_6K_4)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 3,
+                    fieldXPos + noteWidth * 3,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -219,9 +219,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_6K_5)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 4,
+                    fieldXPos + noteWidth * 4,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
@@ -229,9 +229,9 @@ void LaneKeyPressRender(int laneCount) {
 
             if (IsKeyDown(LANE_6K_6)) {
                 DrawRectangle(
-                    (int) fieldXPos + noteWidth * 5,
+                    fieldXPos + noteWidth * 5,
                     0,
-                    (int) noteWidth,
+                    noteWidth,
                     screenHeight,
                     LANE_KEY_PRESS_COLOR
                 );
